Split util
move common
